move getColumns to utils
